+++ spec/lib/sengiri/model/base_spec.rb
@@ -1,43 +1,5 @@
 require 'spec_helper'
 
-ENV['SENGIRI_ENV'] = 'rails_env'
-
-class SengiriModel < Sengiri::Model::Base
-  sharding_group 'sengiri', confs: {
-      'sengiri_shard_1_rails_env'=> {
-        adapter: "sqlite3",
-        database: "spec/db/sengiri_shard_1.sqlite3",
-        pool: 5,
-        timeout: 5000,
-      },
-      'sengiri_shard_second_rails_env'=> {
-        adapter: "sqlite3",
-        database: "spec/db/sengiri_shard_2.sqlite3",
-        pool: 5,
-        timeout: 5000,
-      },
-    }
-end
-
-class SengiriModelWithSuffix < Sengiri::Model::Base
-  sharding_group 'sengiri', {
-    confs: {
-      'sengiri_shard_1_rails_env_suffix'=> {
-        adapter: "sqlite3",
-        database: "spec/db/sengiri_shard_secondary_1.sqlite3",
-        pool: 5,
-        timeout: 5000,
-      },
-      'sengiri_shard_second_rails_env_suffix'=> {
-        adapter: "sqlite3",
-        database: "spec/db/sengiri_shard_secondary_2.sqlite3",
-        pool: 5,
-        timeout: 5000,
-      },
-    },
-    suffix: 'suffix' }
-end
-
 describe SengiriModel do
   # clear db
   before do
