+++ spec/spec_helper.rb
@@ -6,20 +6,8 @@ require File.expand_path('../../lib/sengiri/model/base', __FILE__)
 require "sqlite3"
 require 'pry-byebug'
 
-def setup_test_database(path)
-  # Open a database
-  db = SQLite3::Database.new path
+Dir[File.expand_path("../support/**/*.rb", __FILE__)].each { |f| require f }
 
-  # Create a database
-  rows = db.execute <<-SQL
-    drop table if exists sengiri_models;
-  SQL
-  rows = db.execute <<-SQL
-    create table sengiri_models ( id int PRIMARY KEY, name varchar(30));
-  SQL
+RSpec.configure do |c|
+  c.include Helpers
 end
-
-setup_test_database "spec/db/sengiri_shard_1.sqlite3"
-setup_test_database "spec/db/sengiri_shard_2.sqlite3"
-setup_test_database "spec/db/sengiri_shard_secondary_1.sqlite3"
-setup_test_database "spec/db/sengiri_shard_secondary_2.sqlite3"
